refactor(folder): common > interfaces folder name
