+++ src/common/errors/interfaces/error.interface.ts
